pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bookmark-brain"
version = "0.1.0"
description = "Local-first knowledge retrieval from bookmarks"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "openai>=1.40",
    "trafilatura>=1.12",
    "pyperclip>=1.9",
    "python-frontmatter>=1.1",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.scripts]
bb = "bb.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["bb*"]
